do not use declare(ticks = 1) to catch kill signal

+++ incl/incl.php
@@ -207,31 +207,36 @@ function TimeDiff($time, $opt = array())
     return $str;
 }
 
-$caughtKill = false;
-function CatchKill()
+function CatchKill($use = true)
 {
+    static $caughtKill = false;
     static $setCatch = false;
-    if ($setCatch) {
-        return;
-    }
-    $setCatch = true;
 
     if (PHP_SAPI != 'cli') {
         DebugMessage('Cannot catch kill if not CLI', E_USER_WARNING);
-        return;
+        return false;
     }
 
-    declare(ticks = 1);
-    pcntl_signal(SIGTERM, 'KillSigHandler');
-}
-
-function KillSigHandler($sig)
-{
-    global $caughtKill;
-    if ($sig == SIGTERM) {
-        $caughtKill = true;
-        DebugMessage('Caught kill message, exiting soon..');
+    if (!$use) {
+        $setCatch = $caughtKill = false;
+        return pcntl_signal(SIGTERM, SIG_DFL);
     }
+
+    if ($setCatch) {
+        pcntl_signal_dispatch();
+        return $caughtKill;
+    }
+
+    $setCatch = true;
+
+    pcntl_signal(SIGTERM, function($sig) use (&$caughtKill) {
+        if ($sig == SIGTERM) {
+            $caughtKill = true;
+            DebugMessage('Caught kill message, exiting soon..');
+        }
+    });
+
+    return false;
 }
 
 function RunMeNTimes($howMany = 1)
